+++ ChangeLog.md
@@ -1,5 +1,17 @@
-#0.7.0
+#0.6.1
 
+  * Added int variable "System stations" with the number of stations (of all types) in the system
+  * Added int variable "System starports" with the number of starports in the system
+  * Added int variable "System outposts" with the number of outposts in the system
+  * Added int variable "System planetary stations" with the number of planetary stations (of all types) in the system
+  * Added int variable "System planetary outposts" with the number of planetary outposts in the system
+  * Added int variable "System planetary ports" with the number of planetary ports in the system
+  * Added int variable "Last system stations" with the number of stations (of all types) in the last system
+  * Added int variable "Last system starports" with the number of starports in the last system
+  * Added int variable "Last system outposts" with the number of outposts in the last system
+  * Added int variable "Last system planetary stations" with the number of planetary stations (of all types) in the last system
+  * Added int variable "Last system planetary outposts" with the number of planetary outposts in the last system
+  * Added int variable "Last system planetary ports" with the number of planetary ports in the last system
   * Added string variable 'Ship size' containing the size of the ship (Small, Medium or Large)
 
 #0.6.0